+++ src/google-logger.mjs
@@ -13,7 +13,6 @@ import { Logger } from './logger.mjs';
 import {
   cleanurl, getMaskedTime, getMaskedUserAgent, getSubsystem, isReasonableWeight, isValidCheckpoint,
 } from './utils.mjs';
-import { classifyAcquisition } from './acquisition.mjs';
 
 export class GoogleLogger {
   constructor(req) {
@@ -30,16 +29,6 @@ export class GoogleLogger {
     }
     console.log('logging to Google');
     const now = getMaskedTime(timePadding);
-
-    if (checkpoint === 'utm' && (source === 'utm_source' || source === 'utm_medium')) {
-      /* eslint-disable no-param-reassign */
-      checkpoint = 'acquisition';
-      source = classifyAcquisition(target);
-    }
-    if (checkpoint === 'paid' || checkpoint === 'email') {
-      checkpoint = 'acquisition';
-      source = classifyAcquisition(source, true);
-    }
 
     const data = {
       time: now,